# openapi_billiard.yaml
openapi: 3.0.3
info:
  title: Billiard Club Management System API
  description: Comprehensive API for managing billiard clubs, tables, employees, and billing
  version: 1.0.0
  contact:
    name: Support Team
    email: support@billardmanagement.com
  license:
    name: MIT
servers:
  - url: https://api.billardmanagement.com/v1
    description: Production Server
  - url: http://localhost:3000/v1
    description: Development Server
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Login and receive JWT tokens
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
        '400':
          description: Bad request
  /auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token

  # ==================== CUSTOMERS (SHOP OWNERS) ====================
...
  /bills/{billId}:
    get:
      summary: Get bill details
      tags:
        - Bills
      parameters:
        - in: path
          name: billId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
  /bills/{billId}/pay:
    post:
      summary: Mark bill as paid
      tags:
        - Bills
      parameters:
        - in: path
          name: billId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProcess'
      responses:
        '200':
          description: Bill paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'

  # ==================== BILL DETAILS ====================
  /bills/{billId}/details:
    post:
      summary: Add product to bill
      tags:
        - BillDetails
      parameters:
        - in: path
          name: billId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillDetailCreate'
      responses:
        '201':
          description: Item added to bill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillDetail'

    get:
      summary: Get bill details
      tags:
        - BillDetails
      parameters:
        - in: path
          name: billId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of bill details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BillDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user_role:
          type: string
          enum: [ADMIN, CUSTOMER, EMPLOYEE]
        customer_id:
          type: integer
...
    Bill:
      type: object
      properties:
        bill_id:
          type: integer
        club_id:
          type: integer
        table_id:
          type: integer
        employee_id:
          type: integer
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        total_hours:
          type: number
          format: decimal
        total_table_cost:
          type: number
          format: decimal
        total_product_cost:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        final_amount:
          type: number
          format: decimal
        payment_method:
          type: string
        bill_status:
          type: string
          enum: [Paid, Unpaid, Cancelled]
        created_date:
          type: string
          format: date-time
    PaymentProcess:
      type: object
      required:
        - payment_method
        - amount
      properties:
        payment_method:
          type: string
          enum: [Cash, Card, Transfer]
        amount:
          type: number
          format: decimal
    BillDetailCreate:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
    BillDetail:
      type: object
      properties:
        bill_detail_id:
          type: integer
        bill_id:
          type: integer
        product_id:
          type: integer
        quantity:
          type: integer
        unit_price:
          type: number
          format: decimal
        sub_total:
          type: number
          format: decimal
...